Copy in VPH pre-0.10.18 observers used as superclasses (#49)
- *Category*: observers
- *JIRA issue*: [MIC-3058](https://jira.ihme.washington.edu/browse/MIC-3058)
- *Research reference*: n/a

In preparation for the birth observer and LBWSG integration, we needed to update to vivarium_public_health (VPH) v0.10.18. In order to do that, two subclassed observer classes needed to be added as scaffolding to allow the Mortality and Disability observers to continue to work. Later, if time permits, the Mortality and Disability observers in the latest VPH versions.

Verification and Testing
Ran a `simulate run`. Results were as expected.